+++ src/App.tsx
@@ -1,42 +1,35 @@
 import { Global, ThemeProvider } from '@emotion/react';
 import { GlobalStyle, theme } from './style';
-import Button from './component/@common/Button';
-import Dropdown from './component/@common/Dropdown';
-import { useState } from 'react';
-import ArrowIcon from './component/@common/ArrowIcon';
+import ShoppingList from './page/ShoppingList';
 
 export type SortOption = '높은 가격순' | '낮은 가격순';
 
 function App() {
-  const [selected, setSelected] = useState<SortOption>('낮은 가격순');
+  // const [data, setData] = useState([]);
 
-  const handleSortClick = (content: SortOption) => {
-    setSelected(content);
-  };
-
-  const sortOptions: SortOption[] = ['높은 가격순', '낮은 가격순'];
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  //     const response = await fetch(
+  //       `${import.meta.env.VITE_API_BASE_URL}/cart-items?page=0`,
+  //       {
+  //         headers: {
+  //           'Content-Type': 'application/json',
+  //           Authorization: `Basic ${import.meta.env.VITE_API_KEY}`,
+  //         },
+  //       }
+  //     );
+  //     const results = await response.json();
+  //     console.log('results');
+  //     setData(results);
+  //   };
+  //   fetchData();
+  // }, []);
 
   return (
     <>
       <ThemeProvider theme={theme}>
         <Global styles={GlobalStyle} />
-        <Button variant="default">담기</Button>
-        <Button variant="gray">담기</Button>
-        <Dropdown.Root>
-          <Dropdown.Trigger>
-            {selected}
-            <ArrowIcon />
-          </Dropdown.Trigger>
-          <Dropdown.List>
-            {sortOptions.map((option) => (
-              <Dropdown.Item
-                key={option}
-                handleClick={handleSortClick}
-                content={option}
-              />
-            ))}
-          </Dropdown.List>
-        </Dropdown.Root>
+        <ShoppingList />
       </ThemeProvider>
     </>
   );
